+++ src/lib/components/atoms/divs/FullscreenDiv.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  let { children } = $props();
+</script>
+
+<div class="flex flex-grow flex-col justify-between px-4">
+  {@render children?.()}
+</div>
