+++ src/lib/components/atoms/divs/index.ts
@@ -1,2 +1,3 @@
 export { default as AdaptiveDiv } from "./AdaptiveDiv.svelte";
 export { default as BottomDiv } from "./BottomDiv.svelte";
+export { default as FullscreenDiv } from "./FullscreenDiv.svelte";
