+++ src/lib/components/index.ts
@@ -1,2 +1,1 @@
 export { default as BottomSheet } from "./BottomSheet.svelte";
-export { default as TopBar } from "./TopBar.svelte";
