+++ src/lib/components/molecules/index.ts
@@ -1,2 +1,3 @@
 export * from "./ActionModal.svelte";
 export { default as ActionModal } from "./ActionModal.svelte";
+export { default as TopBar } from "./TopBar.svelte";
